clear code time

+++ pkg/logic/src/time/comp/timer_task_comp.h
@@ -30,7 +30,7 @@ public:
 
     bool IsActive() const;
 
-    int32_t GetEndTime();
+    uint64_t GetEndTime();
     void UpdateEndStamp();
 
     void SetCallBack(const TimerCallback& cb);
